# Add insert hints for each writebatch