--- a/db/memtable.h
+++ b/db/memtable.h
@@ -173,7 +173,8 @@ class MemTable {
   // the <key, seq> already exists.
   bool Add(SequenceNumber seq, ValueType type, const Slice& key,
            const Slice& value, bool allow_concurrent = false,
-           MemTablePostProcessInfo* post_process_info = nullptr);
+           MemTablePostProcessInfo* post_process_info = nullptr,
+           void **hint = nullptr);
 
   // Used to Get value associated with key or Get Merge Operands associated
   // with key.