@Jing118 2019-08-24
Here is the bench result for fillrandom (qps):

`./db_bench --benchmarks=fillrandom -allow_concurrent_memtable_write=true -num=4000000 -batch-size=1 -threads=1 -db=/data3/ylj/tmp -write_buffer_size=536870912 -num_column_families=4`

master:

batch size \ thread num | 1 | 2 | 4 | 8
-- | -- | -- | -- | --
1 | 214724 | 208672 | 272630 | 408532  
10 | 397511 | 333317 | 547891 | 938795  
100 | 439467 | 346279 | 581659 | 1037327 
1000 | 473877 | 378207 | 648497 | 1152165 

with hint_per_batch set to true :

batch size \ thread num | 1 | 2 | 4 | 8
-- | -- | -- | -- | --
1 | 196489 | 210724 | 264376 | 383852  
10 | 358335 | 262496 | 495732 | 917832  
100 | 410783 | 334563 | 561143 | 964826  
1000 | 446459 | 402048 | 625963 | 1088148